Moving files to their correct directories
